Move UnderlyingType to Traits.hpp
