+++ src/Core/include/OpenLoco/Core/Numeric.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "Traits.hpp"
 #include <cstddef>
 #include <cstdint>
 #include <limits>
@@ -11,27 +12,11 @@ namespace OpenLoco::Utility
 
     int32_t bitScanReverse(uint32_t source);
 
-    namespace Detail
-    {
-        // TODO: Move this into Core/Traits.hpp, however Core depends on Utility so we can't do that yet.
-        template<typename T, typename = void>
-        struct UnderlyingType
-        {
-            using type = T;
-        };
-
-        template<typename T>
-        struct UnderlyingType<T, std::enable_if_t<std::is_enum_v<T>>>
-        {
-            using type = typename std::underlying_type<T>::type;
-        };
-    }
-
     template<typename _UIntType>
     static constexpr _UIntType rol(_UIntType x, size_t shift)
     {
         // C++20 replace with std::rotl
-        using T = typename Detail::UnderlyingType<_UIntType>::type;
+        using T = typename Traits::UnderlyingType<_UIntType>::type;
         static_assert(std::is_unsigned_v<T>, "result_type must be an unsigned integral type");
         using limits = typename std::numeric_limits<T>;
         const auto amount = shift & (limits::digits - 1);
@@ -47,7 +32,7 @@ namespace OpenLoco::Utility
     static constexpr _UIntType ror(_UIntType x, size_t shift)
     {
         // C++20 replace with std::rotr
-        using T = typename Detail::UnderlyingType<_UIntType>::type;
+        using T = typename Traits::UnderlyingType<_UIntType>::type;
         static_assert(std::is_unsigned_v<T>, "result_type must be an unsigned integral type");
         using limits = std::numeric_limits<T>;
         const auto amount = shift & (limits::digits - 1);
