+++ src/Core/include/OpenLoco/Core/Traits.hpp
@@ -10,4 +10,17 @@ namespace OpenLoco::Traits
     {
         static constexpr auto value = std::is_trivially_copyable_v<T>;
     };
+
+    // TODO: Move this into Core/Traits.hpp, however Core depends on Utility so we can't do that yet.
+    template<typename T, typename = void>
+    struct UnderlyingType
+    {
+        using type = T;
+    };
+
+    template<typename T>
+    struct UnderlyingType<T, std::enable_if_t<std::is_enum_v<T>>>
+    {
+        using type = typename std::underlying_type<T>::type;
+    };
 }
